AppError setup and error code show

+++ src/app/modules/tasks/taskSerivices.ts
@@ -1,4 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import { StatusCodes } from "http-status-codes"
+import AppError from "../../utils/AppError"
 import { taskInterface } from "./taskInterface"
 import { taskModel } from "./taskSchemaModel"
 
@@ -11,14 +13,26 @@ const getAllTaskFromDb = async () => {
     return result
 }
 const getSignleTaskFromDb = async (id: string) => {
+    const task = await taskModel.findById(id)
+    if(!task){
+        throw new AppError(StatusCodes.NOT_FOUND, "This task is not found")
+    }
     const result = await taskModel.findById(id)
     return result
 }
 const updateSignleTaskFromDb = async (id: string, body: any) => {
+    const task = await taskModel.findById(id)
+    if(!task){
+        throw new AppError(StatusCodes.NOT_FOUND, "This task is not found")
+    }
     const result = await taskModel.findByIdAndUpdate(id, body, {new: true})
     return result
 }
 const deleteSignleTaskFromDb = async (id: string) => {
+    const task = await taskModel.findById(id)
+    if(!task){
+        throw new AppError(StatusCodes.NOT_FOUND,"This task is not found")
+    }
     const result = await taskModel.findByIdAndDelete(id)
     return result
 }
